+++ SeraServer/src/main/java/com/cdtp/seraserver/services/SeraService.java
@@ -17,6 +17,6 @@ public interface SeraService {
     public void addCommand(String employee, String valueName,String value);
     public void addCommandToAllEmployees(String valueName,String value);
     public Map<String,String> getCommand(String employee) throws GreenHouseNotFoundException;
-    public void controlConnection(int total);
+    public void controlConnection(float total);
     public String getIpFromName(String name);
 }
